+++ public/index.php
@@ -2,8 +2,12 @@
 include $_SERVER['DOCUMENT_ROOT'] . '/../services/Autoload.php';
 spl_autoload_register([new Autoload(), 'loadClass']);
 
-echo '<a href="/?c=user&a=user&id=1">Вывести пользователя 1</a>';
-echo '<a href="/?c=user&a=user&id=2">Вывести пользователя 2</a>';
+echo '<a href="/?c=user&a=user&id=1">Вывести пользователя 1 | </a>';
+echo '<a href="/?c=user&a=user&id=2">Вывести пользователя 2 | </a>';
+echo '<a href="/?c=user&a=user&id=2">Вывести всех пользователей | </a>';
+echo '<a href="/?c=user&a=user&id=2">Вывести все товары | </a>';
+echo '<a href="/?c=user&a=user&id=2">Вывести товар 1 | </a>';
+echo '<a href="/?c=user&a=user&id=2">Вывести товар 2 | </a>';
 
 $controllerName = $_GET['c'] ?: 'user';
 $action = $_GET['a'];
